+++ src/main/java/racingcar/view/InputView.java
@@ -2,8 +2,8 @@ package racingcar.view;
 
 import racingcar.view.console.ConsoleReader;
 import racingcar.view.console.ConsoleWriter;
-import racingcar.view.global.exception.CustomException;
-import racingcar.view.global.exception.ErrorMessage;
+import racingcar.global.exception.CustomException;
+import racingcar.global.exception.ErrorMessage;
 
 public class InputView {
     private static final String REQUEST_CARS_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
